Add modified jsonl.gz -> csv conversion script from openalex
The purpose of this script is to convert the data to a format that
can easily be copied into a database. OpenAlex provides a python script
that does this, but takes a long time to run.
I have modified it by running some of the steps concurrently. It still
takes quite a long time.